move a bunch of stuff to terrain
