+++ index.html
@@ -14,7 +14,7 @@
     <script src="./lib/loop.js"></script-->
     <script type="module">
       import * as frame from "./lib/frame.js";
-      
+      frame.draw();
     </script>
   </body>
 </html>
