+++ lib/terrain.js
@@ -1,3 +1,5 @@
+import * as lib from "./lib.js";
+
 export var chunks = [];
 export var chunkIndex = [];
 
@@ -42,29 +44,93 @@ export function drawChunk(data, x, y) {
   }
 }
 
-export function drawTile(tile, x, y) {
+export function drawTile(tile, x, y, can2d, mono) {
   //scale/=4;
-  if (graphics.mono) {
+  if (mono) {
     can2d.fillStyle = tile.color1;
-    can2d.fillRect(x,y,1,1);
+    can2d.fillRect(x*4, y*4,4,4);
   } else {
     for (let i=0; i<16; i++) {
-      if (getBit(tile.pattern, i)) {
+      if (lib.getBit(tile.pattern, i)) {
         can2d.fillStyle = tile.color2;
       } else {
         can2d.fillStyle = tile.color1;
       }
-      can2d.fillRect(x+pmod(i,4)/4,y+(fdiv(i,4))/4,1/4,1/4);
+      can2d.fillRect(x*4+lib.pmod(i,4),y*4+(lib.fdiv(i,4)),1,1);
     }
   }
 }
 
-const tiles = [
+//generation funcs
+
+function serp(x1, x2, a) {return (Math.sin((a-0.5)*Math.PI)/2+0.5)*(x2-x1)+x1}//{return a*(x2-x1)+x1}//
+function serp2d(x1, x2, x3, x4, px, py) {return serp(serp(x1, x2, px), serp(x3, x4, px), py)}
+function serp3d(x1, x2, x3, x4, x5, x6, x7, x8, px, py, pz) {return serp(serp2d(x1, x2, x3, x4, px, py), serp2d(x5, x6, x7, x8, px, py), pz)}
+
+function perlinStep2d(x,y,seed,scale) {
+  return serp2d(
+    hash(Math.floor(x/scale)*scale, Math.floor(y/scale)*scale, seed),
+    hash(Math.ceil(x/scale)*scale, Math.floor(y/scale)*scale, seed),
+    hash(Math.floor(x/scale)*scale, Math.ceil(y/scale)*scale, seed),
+    hash(Math.ceil(x/scale)*scale, Math.ceil(y/scale)*scale, seed),
+    pmod(x/scale, 1),
+    pmod(y/scale, 1)
+  );
+}
+
+function perlin2d(x, y, seed, startScale, endScale) {
+  startScale=2**startScale;
+  endScale=2**endScale;
+  let max=0;
+  let curr=0;
+  for (let i=startScale; i<endScale; i*=2) {
+    max+=i;
+    curr+=perlinStep2d(x, y, seed, i) * i;
+  }
+  
+  return curr/max;
+}
+
+function perlinChunkStep2d(x, y, width, height, seed, scale) {
+  let out = new Array(width*height);
+  x=Math.floor(x)
+  y=Math.floor(y)
+  let ff;
+  let cf;
+  let fc;
+  let cc;
+  for (let j=0; j<height; j++) {
+    let py = y+j;
+    ff = hash(Math.floor(x/scale)*scale, Math.floor(py/scale)*scale, seed);
+    if (pmod(x,scale)==0) cf = ff; else cf = hash(Math.ceil(x/scale)*scale, Math.floor(py/scale)*scale, seed);
+    if (pmod(py,scale)==0) fc = ff; else fc = hash(Math.floor(x/scale)*scale, Math.ceil(py/scale)*scale, seed);
+    if (pmod(x,scale)==0) cc = fc; else cc = hash(Math.ceil(x/scale)*scale, Math.ceil(py/scale)*scale, seed);
+    for (let k=0; k<width; k++) {
+      let px = x+k; // ==; <- hammer
+      //*
+      if (pmod(px,scale)<1) { //problem child
+        ff=cf;
+        fc=cc;
+        cf=hash(Math.ceil(px/scale)*scale+scale, Math.floor(py/scale)*scale, seed);
+        if (pmod(py,scale)==0) cc = cf; else cc = hash(Math.ceil(px/scale)*scale+scale, Math.ceil(py/scale)*scale, seed);
+      }//*/
+      //out[k+j*width] = ff;
+      out[k+j*width] = serp2d(ff,cf,fc,cc,pmod(px,scale)/scale,pmod(py,scale)/scale);
+    }
+  }
+  return out;
+}
+
+function perlinChunk2d(x, y, width, height, seed, startScale, endScale) {
+  
+}
+
+export const tiles = [
   {"color1":"#0000","color2":"#eee1","pattern":"0x342","name":"air"},
   {"color1":"#fff","color2":"#ddd","pattern":"0x3c0","name":"aether"},
   {"color1":"#8f0","color2":"#b60","pattern":"0xfd80","name":"grass"},
   {"color1":"#b60","color2":"#d82","pattern":"0x4020","name":"dirt"},
   {"color1":"#b52","color2":"#a41","pattern":"0x2a8","name":"wood"},
-  {"color1":"#080","color2":"#0000","pattern":"0x2942","name":"wood"},
+  {"color1":"#080","color2":"#0000","pattern":"0x2942","name":"leaves"},
   {"color1":"#48f8","color2":"#2488","pattern":"0x60c","name":"water"},
 ];
